+++ imap_tools/__init__.py
@@ -4,4 +4,4 @@ from .message import *
 from .folder import *
 from .utils import *
 
-__version__ = '0.14.3'
+__version__ = '0.15.0'